Revert #89 - Add IP Location Info
I’ve changed my mind on this, it doesn’t add that much value, and adds
an external hit for that minimal value. There’s the option to pull down
the GeoIP DB to do this locally, but there are issues with the license
and keeping it up to date. May consider adding this again if there’s a
good option, but for now, no.